requirements.txt

Flask
requests
beautifulsoup4
apscheduler
opencv-python
numpy
Pillow
